add basic macro logging and move generate_send_code_for_packet to separate file
